Refactor Aggregate Consumption report to use the new Parameter Adapter, This fixes  - ELMIS-1722: Properly implements Indicator product filter (Which was never implemented) - ELMIS-1460: Remove duplicate products that appears when a product name has changed over time.

+++ modules/report/src/main/java/org/openlmis/report/model/params/BaseParam.java
@@ -23,12 +23,12 @@ import java.util.Date;
 @AllArgsConstructor
 public class BaseParam  {
 
-    private int userId;
+    private Long userId;
 
-    private int page;
+    private Long page;
 
-    private int pageSize;
+    private Long pageSize;
 
-    private int max;
+    private Long max;
 
 }
